+++ internal/provider/provider.go
@@ -119,6 +119,7 @@ func (p *GigahostProvider) Resources(_ context.Context) []func() resource.Resour
 func (p *GigahostProvider) DataSources(_ context.Context) []func() datasource.DataSource {
 	return []func() datasource.DataSource{
 		NewAccountDataSource,
+		NewSSHKeysDataSource,
 	}
 }